+++ src/bin/tectonic/v2cli/commands/mod.rs
@@ -1,0 +1,6 @@
+pub mod build;
+pub mod bundle;
+pub mod dump;
+pub mod new;
+pub mod show;
+pub mod watch;
